refactor: move `extractHashtags()` to `utils` as `hashtags.ts`.
It is a hashtag lexer: `string -> string[]`, with no knowledge of schemas or of
the data model, and `data-model` never called it. Its two consumers (`runner`'s
`wish` builtin and the schema generator) both sit above `utils`, which is
dependency-free and outside the layer stack.

This drops the schema generator's data-model dependency entirely: `schema-tags`
was its only import from that package.

The doc comment now records what the lexer actually does, which is looser than
the convention its syntax implies: no preceding-boundary check (so prose that
merely mentions a tag mints one, as does `#ff0000` or a URL fragment), no
Unicode normalization (so NFD and NFC spellings of the same word never match),
and `\p{Cf}` terminating a token (so Persian is cut at a ZWNJ). Behavior is
unchanged here; the note is a description, not a fix.

`tagsFromSchema()` stays put for now, pending a decision on its home.
